Move AggregationJobCreator to the library target
